Created separate file for framekwork flag and cleaned up original file
